Update firebase library
