+++ ffc/build.gradle
@@ -111,7 +111,7 @@ dependencies {
     implementation "com.android.support:gridlayout-v7:$appcompatVersion"
     implementation "com.android.support:preference-v7:$appcompatVersion"
     implementation 'com.android.support:multidex:1.0.3'
-    implementation 'com.crashlytics.sdk.android:crashlytics:2.9.9'
+    implementation 'com.crashlytics.sdk.android:crashlytics:2.10.0'
     implementation 'com.fatboyindustrial.gson-jodatime-serialisers:gson-jodatime-serialisers:1.6.0'
     implementation 'com.github.bumptech.glide:glide:4.7.1'
     implementation "com.github.chrisbanes:PhotoView:2.1.3"
@@ -127,11 +127,11 @@ dependencies {
     implementation "com.google.android.gms:play-services-maps:16.1.0"
     implementation 'com.google.android.gms:play-services-oss-licenses:16.0.2'
     implementation 'com.google.code.gson:gson:2.8.2'
-    implementation "com.google.firebase:firebase-core:16.0.7"
-    implementation 'com.google.firebase:firebase-auth:16.1.0'
-    implementation 'com.google.firebase:firebase-messaging:17.3.4'
-    implementation 'com.google.firebase:firebase-perf:16.2.3'
-    implementation 'com.google.firebase:firebase-storage:16.0.5'
+    implementation "com.google.firebase:firebase-core:16.0.9"
+    implementation 'com.google.firebase:firebase-auth:17.0.0'
+    implementation 'com.google.firebase:firebase-messaging:18.0.0'
+    implementation 'com.google.firebase:firebase-perf:17.0.0'
+    implementation 'com.google.firebase:firebase-storage:17.0.0'
     implementation 'com.google.maps.android:android-maps-utils:0.5'
     implementation 'com.jakewharton.timber:timber:4.7.1'
     implementation 'com.otaliastudios:zoomlayout:1.3.0'
@@ -140,7 +140,7 @@ dependencies {
     implementation "com.squareup.retrofit2:retrofit:$retrofitVersion"
     implementation 'com.squareup.okhttp3:logging-interceptor:3.10.0'
     implementation 'de.hdodenhof:circleimageview:2.2.0'
-    implementation 'joda-time:joda-time:2.10'
+    implementation 'joda-time:joda-time:2.10.1'
     implementation 'me.piruin:spinney:1.0.0'
     implementation 'me.piruin:phototaker:1.2'
     implementation "org.jetbrains.anko:anko-appcompat-v7-commons:$ankoVersion"
@@ -151,7 +151,7 @@ dependencies {
     implementation "th.or.nectec.android:thai-widget:0.6.2"
 
     testImplementation 'junit:junit:4.12'
-    testImplementation 'org.amshove.kluent:kluent:1.41'
+    testImplementation 'org.amshove.kluent:kluent:1.48'
 
     androidTestImplementation 'org.amshove.kluent:kluent-android:1.40'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
