+++ ffc/src/main/AndroidManifest.xml
@@ -88,12 +88,6 @@
             </intent-filter>
         </service>
 
-        <service android:name=".util.messaging.InstanceIdService">
-            <intent-filter>
-                <action android:name="com.google.firebase.INSTANCE_ID_EVENT"/>
-            </intent-filter>
-        </service>
-
         <activity
             android:name=".MainActivity"
             android:label="@string/title_activity_main"
